Prototype package restructuring
