+++ force_bdss_prototype/__main__.py
@@ -1,4 +1,4 @@
-from force_bdss_prototype.MCOwrapper import MCOwrapper
+from force_bdss_prototype.mco.MCOwrapper import MCOwrapper
 
 A = {"name": "eductA", "manufacturer": "", "pdi": 0}
 B = {"name": "eductB", "manufacturer": "", "pdi": 0}
