+++ force_bdss_prototype/constraints/tests/test_constraints.py
@@ -1,8 +1,9 @@
 import unittest
 import numpy as np
 
-from ..constraints import Constraints
-from ..process_db_access import Process_db_access
+from force_bdss_prototype.constraints.constraints import Constraints
+from force_bdss_prototype.databases.process_db_access import Process_db_access
+
 
 A = { "name": "eductA", "manufacturer": "", "pdi": 0 }
 B = { "name": "eductB", "manufacturer": "", "pdi": 0 }
