+++ force_bdss_prototype/databases/tests/test_material_db_access.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 
-from ..material_db_access import Material_db_access
+from force_bdss_prototype.databases.material_db_access import Material_db_access
 
 A = { "name": "eductA", "manufacturer": "", "pdi": 0 }
 B = { "name": "eductB", "manufacturer": "", "pdi": 0 }
