+++ force_bdss_prototype/mco/MCOwrapper.py
@@ -1,10 +1,10 @@
 import numpy as np
-from .objectives import Objectives
-from .constraints import Constraints
-from .initializer import Initializer
+from ..objectives.objectives import Objectives
+from ..constraints.constraints import Constraints
+from ..initializer.initializer import Initializer
 from .MCOsolver import MCOsolver
-from .pareto_process_db import Pareto_process_db
-from .process_db_access import Process_db_access
+from ..pareto_process.pareto_process_db import Pareto_process_db
+from ..databases.process_db_access import Process_db_access
 import kivy.core.window as window
 from kivy.base import EventLoop
 from kivy.cache import Cache
