+++ force_bdss_prototype/mco/tests/test_MCOwrapper.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 
-from ..MCOwrapper import MCOwrapper
+from force_bdss_prototype.mco.MCOwrapper import MCOwrapper
 
 A = { "name": "eductA", "manufacturer": "", "pdi": 0 }
 B = { "name": "eductB", "manufacturer": "", "pdi": 0 }
